feat: enhance SimpleBreakoutStrategy to accept BacktestConfig and DataKeeper, improve actionType handling in BubbleStorageItem

+++ B3WM.Flutter/lib/models/bubble_storage_item.dart
@@ -23,7 +23,7 @@ class BubbleStorageItem {
       agent: json['agent'] as int,
       amount: (json['amount'] as num).toDouble(),
       date: DateTime.parse(json['date'] as String),
-      actionType: ActionType.fromValue(json['actionType'] as int),
+      actionType: ActionType.fromValue(json['actionType']),
       symbol: json['symbol'] as String,
     );
   }
